+++ stagemonitor-requestmonitor/src/main/java/org/stagemonitor/requestmonitor/RequestLogger.java
@@ -1,16 +1,15 @@
-package org.stagemonitor.requestmonitor.profiler;
+package org.stagemonitor.requestmonitor;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.stagemonitor.requestmonitor.ExecutionContext;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
 
-public class ExecutionContextLogger {
+public class RequestLogger {
 
-	private static final Logger logger = LoggerFactory.getLogger(ExecutionContextLogger.class);
+	private static final Logger logger = LoggerFactory.getLogger(RequestLogger.class);
 
 	private ExecutorService asyncLoggPool = Executors.newSingleThreadExecutor(new ThreadFactory() {
 		@Override
@@ -22,7 +21,7 @@ public class ExecutionContextLogger {
 		}
 	});
 
-	public void logStats(final ExecutionContext requestStats) {
+	public void logStats(final RequestTrace requestTrace) {
 		if (logger.isInfoEnabled()) {
 			asyncLoggPool.execute(new Runnable() {
 				@Override
@@ -30,7 +29,7 @@ public class ExecutionContextLogger {
 					long start = System.currentTimeMillis();
 					StringBuilder log = new StringBuilder(10000);
 					log.append("\n########## PerformanceStats ##########\n");
-					log.append(requestStats.toString());
+					log.append(requestTrace.toString());
 
 					log.append("Printing stats took ").append(System.currentTimeMillis() - start).append(" ms\n");
 					log.append("######################################\n\n\n");