+++ stagemonitor-requestmonitor/src/test/java/org/stagemonitor/requestmonitor/MonitoredMethodExecutionTest.java
@@ -11,9 +11,9 @@ import static org.stagemonitor.core.util.GraphiteSanitizer.sanitizeGraphiteMetri
 
 public class MonitoredMethodExecutionTest {
 
-	private ExecutionContextMonitor.ExecutionInformation<ExecutionContext> executionInformation1;
-	private ExecutionContextMonitor.ExecutionInformation<ExecutionContext> executionInformation2;
-	private ExecutionContextMonitor.ExecutionInformation<ExecutionContext> executionInformation3;
+	private RequestMonitor.RequestInformation<RequestTrace> requestInformation1;
+	private RequestMonitor.RequestInformation<RequestTrace> requestInformation2;
+	private RequestMonitor.RequestInformation<RequestTrace> requestInformation3;
 
 	public void clearState() {
 		getMetricRegistry().removeMatching(new MetricFilter() {
@@ -22,21 +22,21 @@ public class MonitoredMethodExecutionTest {
 				return true;
 			}
 		});
-		executionInformation1 = executionInformation2 = executionInformation3 = null;
+		requestInformation1 = requestInformation2 = requestInformation3 = null;
 	}
 
 	@Test
 	public void testDoubleForwarding() throws Exception {
 		clearState();
-		TestObject testObject = new TestObject(new ExecutionContextMonitor());
+		TestObject testObject = new TestObject(new RequestMonitor());
 		testObject.monitored1();
-		assertEquals(1, executionInformation1.getExecutionResult());
-		assertFalse(executionInformation1.forwardedExecution);
-		assertEquals("monitored1()", executionInformation1.executionContext.getName());
-		assertTrue(executionInformation2.forwardedExecution);
-		assertNull("monitored2()", executionInformation2.executionContext); // forwarded method executions are not monitored
-		assertTrue(executionInformation3.forwardedExecution);
-		assertNull("monitored3()", executionInformation3.executionContext); // forwarded method executions are not monitored
+		assertEquals(1, requestInformation1.getExecutionResult());
+		assertFalse(requestInformation1.forwardedExecution);
+		assertEquals("monitored1()", requestInformation1.request.getName());
+		assertTrue(requestInformation2.forwardedExecution);
+		assertNull("monitored2()", requestInformation2.request); // forwarded method executions are not monitored
+		assertTrue(requestInformation3.forwardedExecution);
+		assertNull("monitored3()", requestInformation3.request); // forwarded method executions are not monitored
 
 		assertNotNull(getMetricRegistry().getTimers().get(name("request", "total", sanitizeGraphiteMetricSegment("monitored1()"))));
 		assertNull(getMetricRegistry().getTimers().get(name("request", "total", sanitizeGraphiteMetricSegment("monitored2()"))));
@@ -47,9 +47,9 @@ public class MonitoredMethodExecutionTest {
 	@Test
 	public void testNormalForwarding() throws Exception {
 		clearState();
-		TestObject testObject = new TestObject(new ExecutionContextMonitor());
+		TestObject testObject = new TestObject(new RequestMonitor());
 		testObject.monitored3();
-		assertEquals(1, executionInformation3.getExecutionResult());
+		assertEquals(1, requestInformation3.getExecutionResult());
 
 		assertNull(getMetricRegistry().getTimers().get(name("request", "total", sanitizeGraphiteMetricSegment("monitored1()"))));
 		assertNull(getMetricRegistry().getTimers().get(name("request", "total", sanitizeGraphiteMetricSegment("monitored2()"))));
@@ -58,44 +58,44 @@ public class MonitoredMethodExecutionTest {
 	}
 
 	private class TestObject {
-		private final ExecutionContextMonitor executionContextMonitor;
+		private final RequestMonitor requestMonitor;
 
-		private TestObject(ExecutionContextMonitor executionContextMonitor) {
-			this.executionContextMonitor = executionContextMonitor;
+		private TestObject(RequestMonitor requestMonitor) {
+			this.requestMonitor = requestMonitor;
 		}
 
 		private int monitored1() throws Exception {
-			executionInformation1 = executionContextMonitor.monitor(
-					new MonitoredMethodExecution("monitored1()", new MonitoredMethodExecution.MethodExecution() {
+			requestInformation1 = requestMonitor.monitor(
+					new MonitoredMethodRequest("monitored1()", new MonitoredMethodRequest.MethodExecution() {
 						@Override
 						public Object execute() throws Exception {
 							return monitored2();
 						}
 					}));
-			return (Integer) executionInformation1.getExecutionResult();
+			return (Integer) requestInformation1.getExecutionResult();
 		}
 
 		private int monitored2() throws Exception {
-			executionInformation2 = executionContextMonitor.monitor(
-					new MonitoredMethodExecution("monitored2()", new MonitoredMethodExecution.MethodExecution() {
+			requestInformation2 = requestMonitor.monitor(
+					new MonitoredMethodRequest("monitored2()", new MonitoredMethodRequest.MethodExecution() {
 						@Override
 						public Object execute() throws Exception {
 							return monitored3();
 						}
 					}));
-			return (Integer) executionInformation2.getExecutionResult();
+			return (Integer) requestInformation2.getExecutionResult();
 		}
 
 		private int monitored3() throws Exception {
-			executionInformation3 = executionContextMonitor.monitor(
-					new MonitoredMethodExecution("monitored3()", new MonitoredMethodExecution.MethodExecution() {
+			requestInformation3 = requestMonitor.monitor(
+					new MonitoredMethodRequest("monitored3()", new MonitoredMethodRequest.MethodExecution() {
 						@Override
 						public Object execute() throws Exception {
 							return notMonitored();
 						}
 					}));
-			System.out.println(executionInformation3);
-			return (Integer) executionInformation3.getExecutionResult();
+			System.out.println(requestInformation3);
+			return (Integer) requestInformation3.getExecutionResult();
 		}
 
 		private int notMonitored() {